requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic
paho-mqtt
